pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "SyntenyQC"
version = "1.0"
requires-python = ">= 3.10"
dependencies = [
  "plotly>=5.16",
  "biopython>=1.80",
  "networkx>=3.1",
  "pandas>=1.5"
]
authors = [{name = "Timothy Kirkwood", email = "tdjkirkwood@hotmail.com"}]
maintainers = [{name = "Timothy Kirkwood", email = "tdjkirkwood@hotmail.com"}]
description = "A command line app for the pre-processing and quality control of synteny plot neighbourhoods"  
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["bgc", "biosynthetic gene cluster", "synteny", "synteny plot", "genome neighbourhood", "neighbourhood"]
classifiers = ["Development Status :: 4 - Beta",
  	       "Programming Language :: Python :: 3",
	       "License :: OSI Approved :: MIT License",
               "Operating System :: Microsoft :: Windows",
               "Intended Audience :: Science/Research",
               "Topic :: Scientific/Engineering :: Bio-Informatics"]

[project.urls]
Homepage = "https://github.com/Tim-Kirkwood/SyntenyQC"
Issues = 'https://github.com/Tim-Kirkwood/SyntenyQC/issues'

[project.scripts]
SyntenyQC = "SyntenyQC.app:main_cli"